+++ state_machine/types/gov.go
@@ -2,6 +2,7 @@ package types
 
 import (
 	"encoding/json"
+	"github.com/ginchuco/ginchu/crypto"
 	"github.com/ginchuco/ginchu/types"
 	"strings"
 )
@@ -44,6 +45,19 @@ func prefixForParamSpace(space string) string {
 	}
 }
 
+func (x *Params) Validate() types.ErrorI {
+	if err := x.Consensus.Validate(); err != nil {
+		return err
+	}
+	if err := x.Fee.Validate(); err != nil {
+		return err
+	}
+	if err := x.Validator.Validate(); err != nil {
+		return err
+	}
+	return x.Governance.Validate()
+}
+
 // consensus param space
 
 const (
@@ -52,6 +66,16 @@ const (
 )
 
 var _ types.ParamSpace = &ConsensusParams{}
+
+func (x *ConsensusParams) Validate() types.ErrorI {
+	if x.BlockSize.Value == 0 {
+		return ErrInvalidParam(ParamBlockSize)
+	}
+	if _, err := x.ParseProtocolVersion(); err != nil {
+		return err
+	}
+	return nil
+}
 
 func (x *ConsensusParams) SetUint64(address string, paramName string, value uint64) types.ErrorI {
 	switch paramName {
@@ -63,7 +87,7 @@ func (x *ConsensusParams) SetUint64(address string, paramName string, value uint
 	default:
 		return ErrUnknownParam()
 	}
-	return nil
+	return x.Validate()
 }
 
 func (x *ConsensusParams) SetOwner(paramName string, owner string) types.ErrorI {
@@ -76,7 +100,7 @@ func (x *ConsensusParams) SetOwner(paramName string, owner string) types.ErrorI 
 	default:
 		return ErrUnknownParam()
 	}
-	return nil
+	return x.Validate()
 }
 
 func (x *ConsensusParams) SetString(address string, paramName string, value string) types.ErrorI {
@@ -92,7 +116,7 @@ func (x *ConsensusParams) SetString(address string, paramName string, value stri
 	default:
 		return ErrUnknownParam()
 	}
-	return nil
+	return x.Validate()
 }
 
 func (x *ConsensusParams) ParseProtocolVersion() (*ProtocolVersion, types.ErrorI) {
@@ -128,19 +152,106 @@ const (
 	ParamValidatorMinStake                  = "validator_min_stake"
 	ParamValidatorMaxCount                  = "validator_max_count"
 	ParamValidatorUnstakingBlocks           = "validator_unstaking_blocks"
-	ParamValidatorMinPauseBlocks            = "validator_min_pause_blocks"
 	ParamValidatorMaxPauseBlocks            = "validator_max_pause_blocks"
 	ParamValidatorMaxEvidenceAgeInBlocks    = "validator_max_evidence_age_in_blocks"
 	ParamValidatorBadProposeSlashPercentage = "validator_bad_propose_slash_percentage"
 	ParamValidatorFaultySignSlashPercentage = "validator_faulty_sign_slash_percentage"
-	ParamValidatorMissedSignSlashPercentage = "validator_missed_sign_slash_percentage"
+	ParamValidatorNonSignSlashPercentage    = "validator_non_sign_slash_percentage"
 	ParamValidatorMaxNonSign                = "validator_max_missed_sign"
 	ParamValidatorNonSignWindow             = "validator_non_sign_window"
 	ParamValidatorDoubleSignSlashPercentage = "validator_double_sign_slash_percentage"
 	ParamValidatorDoubleSignReporterReward  = "validator_double_sign_reporter_reward"
 	ParamValidatorProposerPercentageOfFees  = "validator_proposer_percentage_of_fees"
-	ParamValidatorProposerBlockReward       = "validator_proposer_block_reward"
+	ParamValidatorBlockReward               = "validator_block_reward"
 )
+
+func (x *ValidatorParams) Validate() types.ErrorI {
+	if i, err := types.StringToBigInt(x.ValidatorMinStake.Value); err != nil || types.BigIsZero(i) {
+		return ErrInvalidParam(ParamValidatorMinStake)
+	}
+	if _, err := crypto.NewAddressFromString(x.ValidatorMinStake.Owner); err != nil {
+		return ErrInvalidOwner(ParamValidatorMinStake)
+	}
+	if x.ValidatorMaxCount.Value == 0 {
+		return ErrInvalidParam(ParamValidatorMaxCount)
+	}
+	if _, err := crypto.NewAddressFromString(x.ValidatorMaxCount.Owner); err != nil {
+		return ErrInvalidOwner(ParamValidatorMaxCount)
+	}
+	if x.ValidatorUnstakingBlocks.Value == 0 {
+		return ErrInvalidParam(ParamValidatorUnstakingBlocks)
+	}
+	if _, err := crypto.NewAddressFromString(x.ValidatorUnstakingBlocks.Owner); err != nil {
+		return ErrInvalidOwner(ParamValidatorUnstakingBlocks)
+	}
+	if x.ValidatorMaxPauseBlocks.Value == 0 {
+		return ErrInvalidParam(ParamValidatorMaxPauseBlocks)
+	}
+	if _, err := crypto.NewAddressFromString(x.ValidatorMaxPauseBlocks.Owner); err != nil {
+		return ErrInvalidOwner(ParamValidatorMaxPauseBlocks)
+	}
+	if x.ValidatorMaxEvidenceAgeInBlocks.Value == 0 {
+		return ErrInvalidParam(ParamValidatorMaxEvidenceAgeInBlocks)
+	}
+	if _, err := crypto.NewAddressFromString(x.ValidatorMaxEvidenceAgeInBlocks.Owner); err != nil {
+		return ErrInvalidOwner(ParamValidatorMaxEvidenceAgeInBlocks)
+	}
+	if x.ValidatorBadProposalSlashPercentage.Value > 100 {
+		return ErrInvalidParam(ParamValidatorBadProposeSlashPercentage)
+	}
+	if _, err := crypto.NewAddressFromString(x.ValidatorBadProposalSlashPercentage.Owner); err != nil {
+		return ErrInvalidOwner(ParamValidatorBadProposeSlashPercentage)
+	}
+	if x.ValidatorFaultySignSlashPercentage.Value > 100 {
+		return ErrInvalidParam(ParamValidatorFaultySignSlashPercentage)
+	}
+	if _, err := crypto.NewAddressFromString(x.ValidatorFaultySignSlashPercentage.Owner); err != nil {
+		return ErrInvalidOwner(ParamValidatorFaultySignSlashPercentage)
+	}
+	if x.ValidatorNonSignSlashPercentage.Value > 100 {
+		return ErrInvalidParam(ParamValidatorNonSignSlashPercentage)
+	}
+	if _, err := crypto.NewAddressFromString(x.ValidatorNonSignSlashPercentage.Owner); err != nil {
+		return ErrInvalidOwner(ParamValidatorNonSignSlashPercentage)
+	}
+	if x.ValidatorNonSignWindow.Value == 0 {
+		return ErrInvalidParam(ParamValidatorNonSignWindow)
+	}
+	if _, err := crypto.NewAddressFromString(x.ValidatorNonSignWindow.Owner); err != nil {
+		return ErrInvalidOwner(ParamValidatorNonSignWindow)
+	}
+	if x.ValidatorMaxNonSign.Value < x.ValidatorNonSignWindow.Value {
+		return ErrInvalidParam(ParamValidatorMaxNonSign)
+	}
+	if _, err := crypto.NewAddressFromString(x.ValidatorMaxNonSign.Owner); err != nil {
+		return ErrInvalidOwner(ParamValidatorMaxNonSign)
+	}
+	if x.ValidatorDoubleSignSlashPercentage.Value > 100 {
+		return ErrInvalidParam(ParamValidatorDoubleSignSlashPercentage)
+	}
+	if _, err := crypto.NewAddressFromString(x.ValidatorDoubleSignSlashPercentage.Owner); err != nil {
+		return ErrInvalidOwner(ParamValidatorDoubleSignSlashPercentage)
+	}
+	if _, err := types.StringToBigInt(x.ValidatorDoubleSignReporterReward.Value); err != nil {
+		return ErrInvalidParam(ParamValidatorDoubleSignReporterReward)
+	}
+	if _, err := crypto.NewAddressFromString(x.ValidatorDoubleSignReporterReward.Owner); err != nil {
+		return ErrInvalidOwner(ParamValidatorDoubleSignReporterReward)
+	}
+	if x.ValidatorProposerPercentageOfFees.Value > 100 {
+		return ErrInvalidParam(ParamValidatorProposerPercentageOfFees)
+	}
+	if _, err := crypto.NewAddressFromString(x.ValidatorProposerPercentageOfFees.Owner); err != nil {
+		return ErrInvalidOwner(ParamValidatorProposerPercentageOfFees)
+	}
+	if _, err := types.StringToBigInt(x.ValidatorBlockReward.Value); err != nil {
+		return ErrInvalidParam(ParamValidatorBlockReward)
+	}
+	if _, err := crypto.NewAddressFromString(x.ValidatorBlockReward.Owner); err != nil {
+		return ErrInvalidOwner(ParamValidatorBlockReward)
+	}
+	return nil
+}
 
 func (x *ValidatorParams) SetUint64(address string, paramName string, value uint64) types.ErrorI {
 	switch paramName {
@@ -154,11 +265,6 @@ func (x *ValidatorParams) SetUint64(address string, paramName string, value uint
 			return ErrUnauthorizedParamChange()
 		}
 		x.ValidatorMaxCount.Value = value
-	case ParamValidatorMinPauseBlocks:
-		if address != x.ValidatorMinPauseBlocks.Owner {
-			return ErrUnauthorizedParamChange()
-		}
-		x.ValidatorMinPauseBlocks.Value = value
 	case ParamValidatorMaxPauseBlocks:
 		if address != x.ValidatorMaxPauseBlocks.Owner {
 			return ErrUnauthorizedParamChange()
@@ -189,7 +295,7 @@ func (x *ValidatorParams) SetUint64(address string, paramName string, value uint
 			return ErrUnauthorizedParamChange()
 		}
 		x.ValidatorMaxNonSign.Value = value
-	case ParamValidatorMissedSignSlashPercentage:
+	case ParamValidatorNonSignSlashPercentage:
 		if address != x.ValidatorNonSignSlashPercentage.Owner {
 			return ErrUnauthorizedParamChange()
 		}
@@ -207,7 +313,7 @@ func (x *ValidatorParams) SetUint64(address string, paramName string, value uint
 	default:
 		return ErrUnknownParam()
 	}
-	return nil
+	return x.Validate()
 }
 
 func (x *ValidatorParams) SetString(address string, paramName string, value string) types.ErrorI {
@@ -222,15 +328,15 @@ func (x *ValidatorParams) SetString(address string, paramName string, value stri
 			return ErrUnauthorizedParamChange()
 		}
 		x.ValidatorDoubleSignReporterReward.Value = value
-	case ParamValidatorProposerBlockReward:
-		if address != x.ValidatorProposerBlockReward.Owner {
+	case ParamValidatorBlockReward:
+		if address != x.ValidatorBlockReward.Owner {
 			return ErrUnauthorizedParamChange()
 		}
-		x.ValidatorProposerBlockReward.Value = value
+		x.ValidatorBlockReward.Value = value
 	default:
 		return ErrUnknownParam()
 	}
-	return nil
+	return x.Validate()
 }
 
 func (x *ValidatorParams) SetOwner(paramName string, owner string) types.ErrorI {
@@ -242,8 +348,6 @@ func (x *ValidatorParams) SetOwner(paramName string, owner string) types.ErrorI 
 		x.ValidatorMaxCount.Owner = owner
 	case ParamValidatorUnstakingBlocks:
 		x.ValidatorUnstakingBlocks.Owner = owner
-	case ParamValidatorMinPauseBlocks:
-		x.ValidatorMinPauseBlocks.Owner = owner
 	case ParamValidatorMaxPauseBlocks:
 		x.ValidatorMaxPauseBlocks.Owner = owner
 	case ParamValidatorMaxEvidenceAgeInBlocks:
@@ -256,7 +360,7 @@ func (x *ValidatorParams) SetOwner(paramName string, owner string) types.ErrorI 
 		x.ValidatorBadProposalSlashPercentage.Owner = owner
 	case ParamValidatorFaultySignSlashPercentage:
 		x.ValidatorFaultySignSlashPercentage.Owner = owner
-	case ParamValidatorMissedSignSlashPercentage:
+	case ParamValidatorNonSignSlashPercentage:
 		x.ValidatorNonSignSlashPercentage.Owner = owner
 	case ParamValidatorDoubleSignSlashPercentage:
 		x.ValidatorDoubleSignSlashPercentage.Owner = owner
@@ -264,12 +368,12 @@ func (x *ValidatorParams) SetOwner(paramName string, owner string) types.ErrorI 
 		x.ValidatorDoubleSignReporterReward.Owner = owner
 	case ParamValidatorProposerPercentageOfFees:
 		x.ValidatorProposerPercentageOfFees.Owner = owner
-	case ParamValidatorProposerBlockReward:
-		x.ValidatorProposerBlockReward.Owner = owner
+	case ParamValidatorBlockReward:
+		x.ValidatorBlockReward.Owner = owner
 	default:
 		return ErrUnknownParam()
 	}
-	return nil
+	return x.Validate()
 }
 
 // fee param space
@@ -286,6 +390,58 @@ const (
 	ParamMessageChangeParameterFee = "message_change_parameter_fee"
 	ParamMessageDoubleSignFee      = "message_double_sign_fee"
 )
+
+func (x *FeeParams) Validate() types.ErrorI {
+	if i, err := types.StringToBigInt(x.MessageSendFee.Value); err != nil || types.BigIsZero(i) {
+		return ErrInvalidParam(ParamMessageSendFee)
+	}
+	if _, err := crypto.NewAddressFromString(x.MessageSendFee.Owner); err != nil {
+		return ErrInvalidOwner(ParamMessageSendFee)
+	}
+	if i, err := types.StringToBigInt(x.MessageStakeFee.Value); err != nil || types.BigIsZero(i) {
+		return ErrInvalidParam(ParamMessageStakeFee)
+	}
+	if _, err := crypto.NewAddressFromString(x.MessageStakeFee.Owner); err != nil {
+		return ErrInvalidOwner(ParamMessageStakeFee)
+	}
+	if i, err := types.StringToBigInt(x.MessageEditStakeFee.Value); err != nil || types.BigIsZero(i) {
+		return ErrInvalidParam(ParamMessageEditStakeFee)
+	}
+	if _, err := crypto.NewAddressFromString(x.MessageEditStakeFee.Owner); err != nil {
+		return ErrInvalidOwner(ParamMessageEditStakeFee)
+	}
+	if i, err := types.StringToBigInt(x.MessageUnstakeFee.Value); err != nil || types.BigIsZero(i) {
+		return ErrInvalidParam(ParamMessageUnstakeFee)
+	}
+	if _, err := crypto.NewAddressFromString(x.MessageUnstakeFee.Owner); err != nil {
+		return ErrInvalidOwner(ParamMessageUnstakeFee)
+	}
+	if i, err := types.StringToBigInt(x.MessagePauseFee.Value); err != nil || types.BigIsZero(i) {
+		return ErrInvalidParam(ParamMessagePauseFee)
+	}
+	if _, err := crypto.NewAddressFromString(x.MessagePauseFee.Owner); err != nil {
+		return ErrInvalidOwner(ParamMessagePauseFee)
+	}
+	if i, err := types.StringToBigInt(x.MessageUnpauseFee.Value); err != nil || types.BigIsZero(i) {
+		return ErrInvalidParam(ParamMessageUnpauseFee)
+	}
+	if _, err := crypto.NewAddressFromString(x.MessageUnpauseFee.Owner); err != nil {
+		return ErrInvalidOwner(ParamMessageUnpauseFee)
+	}
+	if i, err := types.StringToBigInt(x.MessageChangeParameterFee.Value); err != nil || types.BigIsZero(i) {
+		return ErrInvalidParam(ParamMessageChangeParameterFee)
+	}
+	if _, err := crypto.NewAddressFromString(x.MessageChangeParameterFee.Owner); err != nil {
+		return ErrInvalidOwner(ParamMessageChangeParameterFee)
+	}
+	if i, err := types.StringToBigInt(x.MessageDoubleSignFee.Value); err != nil || types.BigIsZero(i) {
+		return ErrInvalidParam(ParamMessageDoubleSignFee)
+	}
+	if _, err := crypto.NewAddressFromString(x.MessageDoubleSignFee.Owner); err != nil {
+		return ErrInvalidOwner(ParamMessageDoubleSignFee)
+	}
+	return nil
+}
 
 func (x *FeeParams) SetString(address string, paramName string, value string) types.ErrorI {
 	switch paramName {
@@ -332,7 +488,7 @@ func (x *FeeParams) SetString(address string, paramName string, value string) ty
 	default:
 		return ErrUnknownParam()
 	}
-	return nil
+	return x.Validate()
 }
 
 func (x *FeeParams) SetOwner(paramName string, owner string) types.ErrorI {
@@ -357,7 +513,7 @@ func (x *FeeParams) SetOwner(paramName string, owner string) types.ErrorI {
 	default:
 		return ErrUnknownParam()
 	}
-	return nil
+	return x.Validate()
 }
 
 func (x *FeeParams) SetUint64(address string, paramName string, value uint64) types.ErrorI {
@@ -372,6 +528,13 @@ const (
 
 var _ types.ParamSpace = &GovernanceParams{}
 
+func (x *GovernanceParams) Validate() types.ErrorI {
+	if _, err := crypto.NewAddressFromString(x.AclOwner); err != nil {
+		return ErrInvalidOwner(ParamACLOwner)
+	}
+	return nil
+}
+
 func (x *GovernanceParams) SetOwner(paramName string, owner string) types.ErrorI {
 	switch paramName {
 	case ParamACLOwner:
@@ -379,14 +542,14 @@ func (x *GovernanceParams) SetOwner(paramName string, owner string) types.ErrorI
 	default:
 		return ErrUnknownParam()
 	}
-	return nil
+	return x.Validate()
 }
 
-func (x *GovernanceParams) SetUint64(address string, paramName string, value uint64) types.ErrorI {
+func (x *GovernanceParams) SetUint64(_ string, _ string, _ uint64) types.ErrorI {
 	return ErrUnknownParam()
 }
 
-func (x *GovernanceParams) SetString(address string, paramName string, value string) types.ErrorI {
+func (x *GovernanceParams) SetString(_ string, _ string, _ string) types.ErrorI {
 	return ErrUnknownParam()
 }